lab_2/IS_Lab2_JSON/requirements.txt
PyYAML~=6.0
multipledispatch~=0.6.0